First part of splitting combine_experiments
